refactored more functions to their respective files
