+++ Assets/Scripts/InventoryItem.cs
@@ -7,52 +7,14 @@ using UnityEngine;
 using UnityEngine.Assertions;
 
 
-public class Transaction {
-    public Transaction(float p, float q)
-    {
-        price = p;
-        quantity = q;
-    }
-    public float price;
-    public float quantity;
-}
-public class History : List<Transaction>
-{
-    float min = 0;
-    float max = 0;
-    float avg = 0;
-    public int history_size = 10;
-    ESList prices = new ESList();
-    public float Min() { return min; }
-	public float Max() { return max; }
-    public new void Add(Transaction t)
-    {
-
-        base.Add(t);
-  
-        float sum = 0;
-        int count = 0;
-        min = t.price;
-        max = t.price;
-        for (int i = base.Count-1; i > 0 && i > (base.Count - history_size); i--)
-        {
-            min = Mathf.Min(base[i].price, min);
-            max = Mathf.Max(base[i].price, max);
-            sum += base[i].price;
-            count++;
-        }
-        avg = (count == 0) ? 0 : sum / (float)count;
-
-    }
-}
 public class InventoryItem {
 	public string commodityName { get; private set; }
 	const float significant = 0.25f;
 	const float sig_imbalance = .33f;
 	const float lowInventory = .1f;
 	const float highInventory = 2f;
-	public History buyHistory;
-	public History sellHistory;
+	public TransactionHistory buyHistory;
+	public TransactionHistory sellHistory;
 	public float cost = 1;
     public float last_price = 1;
 	public float wobble = .02f;
@@ -87,8 +49,8 @@ public class InventoryItem {
 	public InventoryItem (string _name, float _quantity=1, float _maxQuantity=10, 
 					float _meanPrice=1, float _production=1)
 	{
-		buyHistory = new History();
-		sellHistory = new History();
+		buyHistory = new TransactionHistory();
+		sellHistory = new TransactionHistory();
 		commodityName = _name;
 		Quantity = _quantity;
 		maxQuantity = _maxQuantity;
